refactor: _document로 파일 이동
